backend/requirements.txt
Flask
Flask-Session
python-dotenv
elevenlabs
ibm_watsonx_ai
openai
